+++ tests/test_train_streams.py
@@ -1,4 +1,3 @@
-import numpy as np
 import os
 import pandas
 import pytest
