+++ wcqtlib/common/utils.py
