+++ wcqtlib/data/parse.py
@@ -43,7 +43,8 @@ import pandas
 import re
 import sys
 
-import wcqtlib.config as C
+import wcqtlib.common.config as C
+import wcqtlib.common.labels
 import wcqtlib.common.utils as utils
 
 if not hasattr(os, 'scandir'):
@@ -491,7 +492,7 @@ def normalize_instrument_names(datasets_df):
         A copy of your dataframe, with instruments only from
         the InstrumentClassMap
     """
-    classmap = InstrumentClassMap()
+    classmap = wcqtlib.common.labels.InstrumentClassMap()
     new_df = datasets_df.copy()
     for i in range(len(new_df)):
         old_class = new_df.iloc[i]["instrument"]
@@ -522,69 +523,7 @@ def load_dataframes(data_dir):
     return result
 
 
-class InstrumentClassMap(object):
-    """Class for handling map between class names and the
-    names they possibly could be from the datasets."""
 
-    def __init__(self, file_path=CLASS_MAP):
-        """
-        Parameters
-        ----------
-        file_path : str
-        """
-        with open(file_path, 'r') as fh:
-            self.data = json.load(fh)
-
-        # Create the reverse map so we can efficiently do the
-        # reverse lookup
-        self.reverse_map = {}
-        for classname in self.data:
-            for item in self.data[classname]:
-                self.reverse_map[item] = classname
-
-        self.index_map = {}
-        for i, classname in enumerate(sorted(self.data.keys())):
-            self.index_map[classname] = i
-
-    @property
-    def allnames(self):
-        """Return a complete list of all class names for searching the
-        dataframe."""
-        return sorted(self.reverse_map.keys())
-
-    @property
-    def classnames(self):
-        return sorted(self.data.keys())
-
-    def __getitem__(self, searchkey):
-        """Get the actual class name. (Actually the reverse map)."""
-        return self.reverse_map.get(searchkey, None)
-
-    def get_index(self, searchkey):
-        """Get the class index for training.
-
-        This is actually the index of the sorted keys.
-
-        Parameters
-        ----------
-        searchkey : str
-
-        Returns
-        -------
-        index : int
-        """
-        return self.index_map[self[searchkey]]
-
-    def from_index(self, index):
-        """Get the instrument name for an index."""
-        return sorted(self.data.keys())[index]
-
-    @property
-    def size(self):
-        """Return the size of the index map (the number of
-        data keys)
-        """
-        return len(self.data.keys())
 
 
 def parse_files_to_dataframe(config):
@@ -670,7 +609,7 @@ def print_stats(config):
             len(inst_filter[inst_filter["dataset"] == "uiowa"]),
             len(inst_filter[inst_filter["dataset"] == "philharmonia"])))
 
-    classmap = InstrumentClassMap()
+    classmap = wcqtlib.common.labels.InstrumentClassMap()
 
     print("---------------------------")
     print("Datasets-Instrument count / dataset")
